add 713

+++ LeetCode/713_乘积小于K的子数组.cpp
@@ -14,7 +14,7 @@ public:
                 prod /= nums[i];
                 i++;
             }
-            result += j - i + 1;
+            result += static_cast<int>(j - i + 1);
         }
         return result;
     }
